multi-restaurant-delivery/xr/server/requirements.txt
fastapi==0.100.0
kafka-python
psycopg2
pytest
